pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robust-portfolio-analytics"
version = "0.1.0"
description = "Robust portfolio analytics and model governance workflow under distribution shift"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy==2.1.3",
  "pandas==2.2.3",
  "scipy==1.14.1",
  "cvxpy==1.5.3",
  "matplotlib==3.9.2",
  "scikit-learn==1.5.2",
  "statsmodels==0.14.4",
  "seaborn==0.13.2",
  "yfinance==0.2.54",
  "jupyterlab==4.3.2",
  "pyarrow==18.1.0",
  "PyYAML==6.0.2",
  "pytest==8.3.4",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
